moved registration test to CreateUser.test from app.test to avoid merging conflicts
